Updated File Serving, Clarified some model fields
As games will need media (pictures) for selection, a Media url has been created for the storage of the pictures required.
This will enable administrators to create games if needed, through forms that upload the pictures.
This feature was suggested by a collaborator.
The file path can be referenced through MEDIA_ROOT.

Some model fields have been clarified, and their type changed to reflect what they will be used for. Main areas of change are DateTimeField's, some to TimeField.
This was done in an effort to reduce effort down the track for development, so we don't have to parse dates to get a time.