Added API models
I have added some models that i believe are required in order to store API details.
Additionally, availability now contains a 'competitive' boolean which states whether or not the user is open for competitive games.
When adding Game_Api_Connection's, you must have an area in the URL that contains '<Player_ID>' so that it can be replaced with the player id when requesting their ranks.

Migrated changes to the database too.